extract Card
