+++ src/components/index.ts
@@ -3,3 +3,4 @@ export * from './MotionStatus';
 export * from './MotionSummary';
 export * from './BatteryLevel';
 export * from './BatterySummary';
+export * from './Card';
